libs/salome/pyproject.toml
[project]
name = "salome"
version = "1.0.0"
requires-python = ">=3.14"
dependencies = [
    "boto3>=1.42.30",
    "discord-interactions>=0.4.0",
    "faiss-cpu>=1.13.2",
    "numpy>=2.4.1",
    "requests>=2.32.5",
    "strands-agents>=1.23.0",
]

[dependency-groups]
dev = [
    "boto3-stubs[bedrock-runtime,s3vectors,sns]>=1.42.30",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/salome"]

